delete-study

+++ src/app/modules/repertoire-menu/repertoire-menu.component.ts
@@ -5,6 +5,7 @@ import { Router } from '@angular/router';
 import { Color } from '../../chess-logic/models';
 import { NewStudyDialogComponent } from '../new-study-dialog/new-study-dialog.component';
 import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog-component';
 
 @Component({
   selector: 'app-repertoire-menu',
@@ -35,9 +36,15 @@ export class RepertoireMenuComponent {
   }
 
   public delete(game: Game): void {
-    this.studyService.deleteStudy(game.studyId).subscribe({
-      complete: () => window.location.reload(),
-      error: (e) => console.log(e)
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {data: {action: `Delete ${game.title}`}});
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.studyService.deleteStudy(game.studyId).subscribe({
+          complete: () => window.location.reload(),
+          error: (e) => console.log(e)
+        });
+      }
     });
   }
